API and module layout cleanup
