+++ pydsfapi/src/pydsfapi/__init__.py
@@ -1,0 +1,4 @@
+SOCKET_DIRECTORY = "/var/run/dsf"
+SOCKET_FILE = "dcs.sock"
+FULL_SOCKET_PATH = SOCKET_DIRECTORY + "/" + SOCKET_FILE
+DEFAULT_BACKLOG = 4
