+++ pydsfapi/src/pydsfapi/connections.py
@@ -17,23 +17,17 @@ connections provides different classes for connections
     You should have received a copy of the GNU Lesser General Public License
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
-from enum import Enum
-import asyncio
 import json
 import os
 import socket
-from concurrent.futures import ThreadPoolExecutor
 from typing import Optional
 
+from . import DEFAULT_BACKLOG, FULL_SOCKET_PATH
 from .commands import responses, basecommands, code, result, codechannel
 from .commands.basecommands import MessageType
 from .initmessages import serverinitmessage, clientinitmessages
-from .model import parsedfileinfo, machinemodel
-
-SOCKET_DIRECTORY = "/var/run/dsf"
-SOCKET_FILE = "dcs.sock"
-FULL_SOCKET_PATH = SOCKET_DIRECTORY + "/" + SOCKET_FILE
-DEFAULT_BACKLOG = 4
+from .http import HttpEndpointUnixSocket
+from .models import MachineModel, ParsedFileInfo
 
 
 class TaskCanceledException(Exception):
@@ -48,180 +42,6 @@ class InternalServerException(Exception):
         self.command = command
         self.error_type = error_type
         self.error_message = error_message
-
-
-class HttpResponseType(str, Enum):
-    """Enumeration of supported HTTP responses"""
-
-    StatusCode = "StatusCode"
-    PlainText = "PlainText"
-    JSON = "JSON"
-    File = "File"
-
-
-class ReceivedHttpRequest:
-    """Notification sent by the webserver when a new HTTP request is received"""
-
-    @classmethod
-    def from_json(cls, data):
-        """Deserialize an instance of this class from deserialized JSON dictionary"""
-        return cls(**data)
-
-    def __init__(
-        self, sessionId: int, queries: dict, headers: dict, contentType: str, body: str
-    ):
-        self.session_id = sessionId
-        self.queries = queries
-        self.headers = headers
-        self.content_type = contentType
-        self.body = body
-
-
-class HttpEndpointConnection:
-    """Connection class for dealing with requests received from a custom HTTP endpoint"""
-
-    def __init__(self, reader, writer, is_websocket: bool, debug: bool = False):
-        """Constructor for a new connection dealing with a single HTTP endpoint request"""
-        self.reader = reader
-        self.writer = writer
-        self.is_websocket = is_websocket
-        self.debug = debug
-
-    def close(self):
-        """Close the connection"""
-        self.writer.close()
-
-    async def read_request(self):
-        """
-        Read information about the last HTTP request.
-        Note that a call to this method may fail!
-        """
-        return await self.receive(ReceivedHttpRequest)
-
-    async def send_response(
-        self,
-        status_code: int = 204,
-        response: str = "",
-        response_type: HttpResponseType = HttpResponseType.StatusCode,
-    ):
-        """
-        Send a simple HTTP response to the client and dispose
-        this connection unless it is a WebSocket.
-        """
-        try:
-            await self.send(
-                {
-                    "StatusCode": status_code,
-                    "Response": response,
-                    "ResponseType": response_type,
-                }
-            )
-        finally:
-            # Close this connection automatically if only one response can be sent
-            if not self.is_websocket:
-                self.close()
-
-    async def receive(self, cls):
-        """Receive a deserialized object"""
-        json_string = await self.receive_json()
-        return cls.from_json(json.loads(json_string))
-
-    async def receive_json(self):
-        """Receive a JSON object"""
-        json_string = (await self.reader.read(32 * 1024)).decode("utf8")
-        if self.debug:
-            print("recv: {0}".format(json_string))
-        return json_string
-
-    async def send(self, obj):
-        """Send an arbitrary object"""
-        json_string = json.dumps(obj, default=lambda o: o.__dict__)
-        if self.debug:
-            print("send: {0}".format(json_string))
-        self.writer.write(json_string.encode("utf8"))
-        await self.writer.drain()
-
-
-class HttpEndpointUnixSocket:
-    """Class for dealing with custom HTTP endpoints"""
-
-    def __init__(
-        self,
-        endpoint_type: basecommands.HttpEndpointType,
-        namespace: str,
-        path: str,
-        socket_path: str,
-        backlog: int = DEFAULT_BACKLOG,
-        debug: bool = False,
-    ):
-        """Open a new UNIX socket on the given file path"""
-        self.endpoint_type = endpoint_type
-        self.namespace = namespace
-        self.endpoint_path = path
-        self.socket_path = socket_path
-        self.backlog = backlog
-        self.handler = None
-        self.debug = debug
-        self._loop = None
-        self._server = None
-
-        try:
-            os.remove(self.socket_path)
-        except FileNotFoundError:
-            # We don't care if the file was missing
-            # TODO: should we care about deletion failed?
-            pass
-
-        self.executor = ThreadPoolExecutor(max_workers=1)
-        self.event_loop = self.executor.submit(self.start_connection_listener)
-
-    def close(self):
-        """Close the socket connection"""
-        if self._loop is not None:
-            # TODO: this enables correctly ending the loop. Why?
-            self._loop.set_debug(True)
-            self._server.close()
-            self._loop.stop()
-        self.event_loop.cancel()
-        self.executor.shutdown(wait=False)
-        try:
-            os.remove(self.socket_path)
-        except FileNotFoundError:
-            pass
-
-    def set_endpoint_handler(self, handler):
-        """Set the handler to handle client connections"""
-        self.handler = handler
-
-    def start_connection_listener(self):
-        try:
-            self._loop = asyncio.new_event_loop()
-            self._server = asyncio.start_unix_server(
-                self.handle_connection, self.socket_path, backlog=self.backlog
-            )
-            self._loop.create_task(self._server)
-            self._loop.run_forever()
-        finally:
-            self._loop.close()
-
-    async def handle_connection(self, reader, writer):
-        """Handle incoming UNIX socket connections (HTTP/WebSocket requests)"""
-        http_endpoint_connection = HttpEndpointConnection(
-            reader,
-            writer,
-            self.endpoint_type == basecommands.HttpEndpointType.WebSocket,
-            debug=self.debug,
-        )
-        if self.handler is not None:
-            # Invoke the event handler and forward the wrapped connection for dealing
-            # with a single endpoint connection.
-            # Note that the event delegate is responsible for disposing the connection!
-            await self.handler(http_endpoint_connection)
-        else:
-            await http_endpoint_connection.send_response(
-                500, "No event handler registered"
-            )
-            http_endpoint_connection.close()
 
 
 class BaseConnection:
@@ -306,8 +126,10 @@ class BaseConnection:
         json_string = self.receive_json()
         return json.loads(json_string, object_hook=responses.decode_response)
 
-    def receive_json(self):
+    def receive_json(self) -> str:
         """Receive the JSON response from the server"""
+        if not self.socket:
+            raise RuntimeError("socket is closed or missing")
 
         json_string = self.input
 
@@ -342,7 +164,7 @@ class BaseConnection:
                     found = True
 
         if self.debug:
-            print("recv: {0}".format(json_string))
+            print("recv:", json_string)
         return json_string
 
     @staticmethod
@@ -412,7 +234,7 @@ class BaseCommandConnection(BaseConnection):
     def get_file_info(self, file_name: str):
         """Parse a G-code file and returns file information about it"""
         res = self.perform_command(
-            basecommands.get_file_info(file_name), parsedfileinfo.ParsedFileInfo
+            basecommands.get_file_info(file_name), ParsedFileInfo
         )
         return res.result
 
@@ -426,9 +248,7 @@ class BaseCommandConnection(BaseConnection):
 
     def get_object_model(self):
         """Retrieve the full object model of the machine."""
-        res = self.perform_command(
-            basecommands.get_object_model(), machinemodel.MachineModel
-        )
+        res = self.perform_command(basecommands.get_object_model(), MachineModel)
         return res.result
 
     def get_serialized_machine_model(self):
@@ -621,7 +441,7 @@ class InterceptConnection(BaseCommandConnection):
 
         return super().connect(iim, socket_path)
 
-    def receive_code(self):
+    def receive_code(self) -> code.Code:
         """Wait for a code to be intercepted and read it"""
         return self.receive(code.Code)
 
@@ -634,7 +454,7 @@ class InterceptConnection(BaseCommandConnection):
         self.send(basecommands.ignore())
 
     def resolve_code(
-        self, rtype: MessageType = MessageType.Success, content: str = None
+        self, rtype: MessageType = MessageType.Success, content: Optional[str] = None
     ):
         """
         Instruct the control server to resolve the last received code with the given
@@ -665,17 +485,17 @@ class SubscribeConnection(BaseConnection):
         )
         return super().connect(sim, socket_path)
 
-    def get_machine_model(self):
+    def get_machine_model(self) -> MachineModel:
         """
         Retrieves the full object model of the machine
         In subscription mode this is the first command that has to be called once a
         ConnectionAbortedError has been established.
         """
-        machine_model = self.receive(machinemodel.MachineModel)
+        machine_model = self.receive(MachineModel)
         self.send(basecommands.acknowledge())
         return machine_model
 
-    def get_serialized_machine_model(self):
+    def get_serialized_machine_model(self) -> str:
         """
         Optimized method to query the machine model UTF-8 JSON in any mode.
         May be used to get machine model patches as well.
@@ -684,7 +504,7 @@ class SubscribeConnection(BaseConnection):
         self.send(basecommands.acknowledge())
         return machine_model_json
 
-    def get_machine_model_patch(self):
+    def get_machine_model_patch(self) -> str:
         """
         Receive a (partial) machine model update.
         If the subscription mode is set to SubscriptionMode.PATCH new update patches of
